docs: note automatic config saving
